add MKV processing
Added a sub-function to check if input file is MKV and if so convert to MP4 before encoding with HandBrake.  File also creates ~/encode.pid to prevent multiple runs so this can be set to run every minute or every 5 minutes.  Upon successful encode the output and time is logged to ~/converted.log

If you add to crontab ensure it is the crontab for the user that runs sickrage or else your files won't be accessible by Sonnar or Sickrage.